obsidian_api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pyyaml==6.0.2
pydantic==2.10.3
